Added recover from panic middleware
Each http connection is its own go routine. When panic happens, we
return from handler functions and on the way out execute each defer
function. With in build function recover() we can check if any panic
happened while handling request. If it did happen we can display nice
error message for user (this function is defered so it will execute in
event of the panic). Difference with using recoverPanic middleware is
that here page don't show couldn't load website logo but shows nice text
message of the error.